Move IGameConfigPtr to a separate header
